Invalid memory access in Client Server examples
The ngtcp2_conn_callbacks and ngtcp2_settings structures that
are passed into ngtcp2_conn_{client,server}_new only exist in
the scope of the stack that the above function is called.

Once the function returns these objects no longer exist yet
they continue to be referenced throughout the lifetime of
the conn.

Fix this by making the callback and settings structures into
members of the Client and (Server) Handler classes.