Add a counter to rtb_entry
Count the number of times a packet has been retransmitted.
At this stage, it's only purpose is to slowly backoff
retransmit timeouts so as not to cause flooding.
In future it could be used for statistics reporting and
telemetry, especially if a similar counter is kept at
the connection level.