Allow handshake callbacks to run asynchronously
When using OpenSSL for TLS and its asynchronous certificate callback
to select a certificate, it can happen that the callback call cannot
be immediately completed, and needs to be re-attempted. In which case
the OpenSSL will return an error like "SSL_X509_LOOKUP" (which is not
a real failure) and the application's ngtcp2 callback need to signal
to ngtcp2 that no work has been done, but there wasn't a failure either.

This adds a new return error value, so that the application that invoked
ngtcp2 can distinguish between the two cases.