Make HTTPHeader struct templated
Currently, the HTTPHeader constructor is templated with the types for
the header name and the value, but the actual name and value members are
of std::string type. It is currently not possible to have a different
type for the value for example.

This commit suggests making the struct templated to allow for values
other than std::string. I'm not sure if that is the way to go or perhaps
it would be acceptable to only allow std::string values, but wanted to
open this up for feedback.