Fix example server's write callback never being called
Handler::wev_ is initialized with a file descriptor set to 0. ev_io_set() is never called for this watcher, so its file descriptor remains 0 and calls to ev_io_start() don't have the desired effect. As a result writecb() is never called.

There may still be an issue due to all endpoints sharing the same watcher, but this is mostly orthogonal to the problem fixed here.