Reclaim 1 frame per probe packet
Reclaiming additional frames might send more frames in one packet, but
it might send another probe packet if they are large frames like
STREAM or CRYPTO.  Even if we just only reclaim one frame, if it is
STREAM or CRYPTO and it does not fit into one packet, it is split into
two, and the latter is sent in additional probe packet, which is not
the intentional behaviour.  But at the moment, there is no mechanism
to stop that.