Remove NGTCP2_PV_FLAG_MTU_PROBE
NGTCP2_PV_FLAG_MTU_PROBE was introduced to deal with the second path
validation due to amplification limit.  If the path validation with
that flag failed, ngtcp2 returns NGTCP2_ERR_NO_VIALBE_PATH.  But we
have always the fallback path when this second path validation is
performed.  We can just fallback to the old path when the 2nd
validation failed, which removes the need to return
NGTCP2_ERR_NO_VIALBE_PATH and NGTCP2_PV_FLAG_MTU_PROBE flag.