Fix compile error with libressl
This change fixes compile errors with libressl.  This is a stopgap
solution.  If the difference between libressl and quictls gets larger,
we will consider to make a separate library for libressl.

Applications in examples do not compile with libressl, and it is a
known fact, and we do not fix it until libressl gets its own
ngtcp2_crypto library.