Loosen app-limited condition
As TCP Linux does, during slow-start, only enter app-limited when less
than half of CWND is used.  Also do not enter app-limited during
handshake because we have not that much to send in that state, and it
would be nicer to warm up CWND with handshake bytes for the upcoming
application traffic.