Remove outdated PMTUD probe packet handling when it is lost
It seems this is an oversight in
cbd6f5815c0c6e8ae8040e1ce255103f3ff3cc6a.  This relevant part of code
should have been removed in that commit.  Because this function is
only called when Retry packet is received, and PMTUD packet is not
sent in this phase, this extra code should not cause any issue.