bbr: Avoid modifying ngtcp2_rs on packet loss
Previously, ngtcp2_rs conn->rst->rs is modified when a packet is lost.
But this is a bit strange, and it is the only place we do in BBR.
This change uses new temporary variable ngtcp2_rs on packet loss
without modifying the existing conn->rst->rs.  With this change, we
observed improvements in the cross traffic test in
quic-interop-runner.