cc: Compute pacing interval and send_quantum
Compute pacing interval and send_quantum as we do in BBR.  Previously
we always set send_quantum to 64k, which is too bursty.  It might be
good for benchmark on loopback but has severe impact on real internet
traffic due to extra burstiness.  The idea is send 1ms worth of data
at once.  This change does not affect 1 connection throughput as long
as event loop keeps up to this 1ms interval.