cc: Increase minimum send_quantum to 10
The previous value of minimum value 2 is too small because usually an
event loop cannot keep up the 1ms tick and UDP send call is too
expensive.  RFC 9002 says that the burst should be limited to the
initial congestion window, which is 10 packets, which suggests that 10
packets burst is acceptable for the traditional loss based congestion
controller with pacing over smoothed RTT and CWND.  10 is still
conservative against full GSO buffer.