examples: Increase number of packets to read
It turns out that the limit of 10 packets per event loop is too small,
that prevents an endpoint from consuming ACKs and other control frames
(e.g., MAX_STREAM_DATA, MAX_STREAMS), resulting in the loss of
throughput.  This change increases maximum number of packets to read
to 64.  Meanwhile, BBR is very sensitive to event loop tick because
its GSO burst is limited to 1ms window.  For BBR, keep reading packets
until 1ms is passed or 64 packets are read.