ngtcp2_map: Optimize a little bit further
Now separate key, data and psl so that they are more cache friendly
because primarily we use psl to probe the sequences.  Previously, data
== NULL means that no element is stored there, but now it is replaced
by psl[i] == 0.  psl is now uint8_t and the maximum probe sequence
length is 254 (because we allocate 0 for special purpose).  The
probability of fail probing 255 times with the load factor 7 / 8 is
vanishingly small (~10**-15), and thus it does not happen in practice
when considering the usage of this hash table in our library.