crypto: Allow an application to store opaque data to a regular token
This change adds ngtcp2_crypto_generate_regular_token2 which can embed
the opaque data to a regular token.
ngtcp2_crypto_verify_regular_token2 can retrieve the opaque data on
successful verification.  Because the token is repeated in every
Initial packet from a client, and the default UDP datagram payload
size is 1200, the size of token is severely limited.  If larger data
is required, consider to store it in a TLS session ticket.  But
storing the opaque data in the TLS session ticket is not widely
implemented as of this writing.  Therefore storing data in QUIC token
is still a viable option if server needs to store some data and get it
back on reconnection.