examples: Add simulator
Simulator consists of 2 endpoints: client and server that are ngtcp2
endpoints and have the real wolfSSL TLS stack.  The network path
between them are emulated in the simulator.  We can change delay,
rate, limit, and loss to simulate very simple network configurations.
Because ngtcp2 takes timestamp as a parameter, the time is also
simulated, and skips to the next event.  This allows us to test a
scenario that takes very long time in real time (e.g., 1 hour RTT).