ngtcp2_ksl: Remove alignment enforcement for keys
The intent to add this alignment is align key data to 8 bytes boundary
so that we can read uint64_t variable by pointer cast.  But it turns
out that aligning keys is useless.

On 64-bit systems, uint8_t *keys is aligned to 8 bytes boundary, then
the actual keys following this field are aligned to 8 bytes too.

On 32-bit systems, uint8_t *keys is aligned to 4 bytes boundary, then
the keys is also aligned to 4 bytes boundary.  On these systems,
64-bit integer is aligned to 4 bytes boundary as well, so no need to
care about the alignment.