Export ngtcp2_crypto_ossl(_static) targets
Using the existing ${PROJECT_NAME}Targets which is
installed by the /lib directory.
Wrap individual ngtcp2_crypto_ossl_INCLUDE_DIRS into
BUILD_INTERFACE generator expressions.
The INSTALL_INTERFACE for includes is identical to the
public ngtcp2 link library, so it can be omitted here.
OpenSSL::SSL is a public link library of ossl crypt,
so it must be looked up with find_dependency in the
installed CMake config file.