frame_chain: Use a flag to determine the allocation type
Previously, we use the size of frame data to determine whether
ngtcp2_frame_chain is allocated by ngtcp2_objalloc.  But this is
fragile and we already have couple of bugs that lead to potential
memory leak.  We also need to write additional code to create new
ngtcp2_frame_chain if the relevant field changes.

In this commit, we introduced a flag to determine the allocation type
for ngtcp2_frame_chain to fix bugs and it also can bring more
optimization opportunities.